Added extended attributes: sign '@'; and code refactoring
